question_6_genai/requirements.txt
pandas>=2.0
pydantic>=2.0
langchain>=0.3
langchain-anthropic>=1.1.0
python-dotenv>=1.0
